Create muncher.py
This project was made by myself and a partner for our final class project. It has a few problems in terms of versatility such as not being able to attack windows. We are going to implement a brute force method later using a served up wordlist via http. For demonstration purposes, we used a hard coded user and password. We'll be using ttl numbers to identify the OS and change the attack method accordingly. I've tested code for ttl numbers and it's promising so far. Please use this code for educational purposes only(as we have done). This code was only executed in controlled environments using virtual machines on virtual networks.